+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.6
+
+- Add `validate(options: { isSubmit: boolean })` method to the Form Object, to trigger form validation without the `onSubmit` action.
+- Add `options: { isSubmit: boolean }` to the validator, to allow varied logic depending on the validation trigger.
+- Fix TypeScript definition for FormErrors
+
 ## 0.3.5
 
 - Add `hasKeyOrAlias` to experimental Form Groups.